# openapi.yaml
components:
  schemas:
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          title: Detail
          type: array
      title: HTTPValidationError
      type: object
    PaginatedResponse:
      properties:
        has_next:
          title: Has Next
          type: boolean
        has_prev:
          title: Has Prev
          type: boolean
        items:
          items: {}
          title: Items
          type: array
        page:
          title: Page
          type: integer
        per_page:
          title: Per Page
          type: integer
        total:
          title: Total
          type: integer
      required:
      - items
      - total
      - page
      - per_page
      - has_next
      - has_prev
      title: PaginatedResponse
      type: object
    SubscriberCreate:
      properties:
        custom_fields:
          default: {}
          title: Custom Fields
          type: object
        email:
          format: email
          title: Email
          type: string
        first_name:
          anyOf:
          - type: string
          - nullable: true
          title: First Name
        last_name:
          anyOf:
          - type: string
          - nullable: true
          title: Last Name
        status:
          allOf:
          - $ref: '#/components/schemas/SubscriberStatus'
          default: active
        tags:
          default: []
          items:
            type: string
          title: Tags
          type: array
        user_id:
          format: uuid
          title: User Id
          type: string
      required:
      - email
      - user_id
      title: SubscriberCreate
      type: object
    SubscriberResponse:
      properties:
        created_at:
          format: date-time
          title: Created At
          type: string
        custom_fields:
          default: {}
          title: Custom Fields
          type: object
        email:
          format: email
          title: Email
          type: string
        first_name:
          anyOf:
          - type: string
          - nullable: true
          title: First Name
        id:
          format: uuid
          title: Id
          type: string
        last_name:
          anyOf:
          - type: string
          - nullable: true
          title: Last Name
        source:
          anyOf:
          - type: string
          - nullable: true
          title: Source
        status:
          allOf:
          - $ref: '#/components/schemas/SubscriberStatus'
          default: active
        tags:
          default: []
          items:
            type: string
          title: Tags
          type: array
        updated_at:
          anyOf:
          - format: date-time
            type: string
          - nullable: true
          title: Updated At
        user_id:
          format: uuid
          title: User Id
          type: string
      required:
      - email
      - id
      - user_id
      - created_at
      title: SubscriberResponse
      type: object
    SubscriberStatus:
      enum:
      - active
      - unsubscribed
      - bounced
      title: SubscriberStatus
      type: string
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          title: Location
          type: array
        msg:
          title: Message
          type: string
        type:
          title: Error Type
          type: string
      required:
      - loc
      - msg
      - type
      title: ValidationError
      type: object
...
  securitySchemes:
    OAuth2PasswordBearer:
      flows:
        password:
          scopes: {}
          tokenUrl: /api/v1/login/access-token
      type: oauth2
info:
  description: API for managing subscribers, integrations, and webhooks
  title: Whookfirm Subscriber Management API
  version: 1.0.0
openapi: 3.1.0
paths:
...
  /api/v1/subscribers:
    get:
      description: List subscribers with pagination and filtering
      operationId: list_subscribers_api_v1_subscribers_get
      parameters:
      - in: query
        name: page
        required: false
        schema:
          default: 1
          minimum: 1
          title: Page
          type: integer
      - in: query
        name: per_page
        required: false
        schema:
          default: 50
          maximum: 100
          minimum: 1
          title: Per Page
          type: integer
      - in: query
        name: search
        required: false
        schema:
          anyOf:
          - type: string
          - nullable: true
          title: Search
      - in: query
        name: status
        required: false
        schema:
          anyOf:
          - $ref: '#/components/schemas/SubscriberStatus'
          - nullable: true
          title: Status
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedResponse'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - OAuth2PasswordBearer: []
      summary: List Subscribers
      tags:
      - subscriber
    post:
      description: Create a new subscriber
      operationId: create_subscriber_api_v1_subscribers_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SubscriberCreate'
        required: true
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubscriberResponse'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Create Subscriber
      tags:
      - subscriber
...